Fix query builder for 2.x
Why:
The build was not function as it supposed to be
due to the use of `_.set(body, 'query.bool.must', query)`
when building a query with filter. It build incorrect
query when the query was a combination, which build
something like `bool: {must:[...]}`. Hence the
end result was:
```
{
query: {
bool: {
filter: {
// some filters
		}
	  },
must: {
	bool: { must: [...] }
	  }
	   }
}
```

This change addresses the need by:
Build query for filters and queries separately
and merge it after that.

This change affects the following: